docs: add comments and update README

+++ index.js
@@ -12,9 +12,9 @@ const crypto = require('crypto');
 app.use(async ctx => {
   const public = config.public;
   if (public === true) {
-    publicMode(ctx)
+    publicMode(ctx);
   } else {
-    privateMode(ctx)
+    privateMode(ctx);
   };
 });
 
@@ -26,6 +26,15 @@ app.listen(port, function() {
   console.log('Server times is', config.times);
 });
 
+/**
+ * Use to generate key and file name in public mode
+ * return `null` when requestKey, device or configKey doesn't exist
+ * return a json when all exist
+ * {
+ *    file: IP storage file path
+ *    key: file name and the secret key
+ * }
+ */
 function generateKey(ctx) {
   const request = ctx.request.query.request;
   const host = config.host;
@@ -49,6 +58,9 @@ function generateKey(ctx) {
   };
 };
 
+/**
+ * Public mode, without matching the config key and the request key
+ */
 function publicMode(ctx) {
   const request = ctx.request.query.request;
   const requestKey = ctx.request.query.key;
@@ -58,7 +70,7 @@ function publicMode(ctx) {
       const dataFile = data.file;
       const key = data.key
       if (request === 'add') {
-        clean(dataFile);
+        clean(ctx, dataFile, false);
         const addData = {
           ctx: ctx,
           dataFile: dataFile,
@@ -73,11 +85,17 @@ function publicMode(ctx) {
   } else if (request === 'get' && requestKey) {
     const dataFile = './data/' + requestKey + '.json';
     get(ctx, dataFile);
+  } else if (request === 'delete' && requestKey) {
+    const dataFile = './data/' + requestKey + '.json';
+    clean(ctx, dataFile, true)
   } else {
     ctx.response.status = 403;
   };
 };
 
+/**
+ * Private mode, with matching the config key and the request key
+ */
 function privateMode(ctx) {
   const request = ctx.request.query.request;
   const requestKey = ctx.request.query.key;
@@ -86,7 +104,7 @@ function privateMode(ctx) {
   if (requestKey === configKey && device) {
     const dataFile = './data/' + device + '.json';
     if (request === 'add') {
-      clean(dataFile);
+      clean(ctx, dataFile, false);
       const addData = {
         ctx: ctx,
         dataFile: dataFile,
@@ -96,24 +114,48 @@ function privateMode(ctx) {
       add(addData);
     } else if (request === 'get') {
       get(ctx, dataFile);
+    } else if (request === 'delete') {
+      clean(ctx, dataFile, true)
     };
   } else {
     ctx.response.status = 403;
   };
 };
 
-function clean(dataFile) {
+/**
+ * Clean data, this can be called by request or when it reach the maximum time
+ * ctx -> koa variable
+ * dataFile -> IP storage file path
+ * request -> if is true, means it requests to be deleted
+ *            if is false, means it reachs the maximum time
+ */
+function clean(ctx, dataFile, request) {
   const obj = jsonfile.readFileSync(dataFile, {throws: false});
   const times = config.times;
   if (obj === null) {
-    return;
+    ctx.response.status = 403;
   } else {
-    if (obj.times === times) {
-      fs.unlinkSync(dataFile)
+   if (request === true) {
+      const succeededJson = {
+        status: "Succeeded"
+      };
+      fs.unlinkSync(dataFile);
+      ctx.body = succeededJson;
+    } else if (obj.times === times) {
+      fs.unlinkSync(dataFile);
     };
   }
 };
 
+/**
+ * Add IP
+ * addData -> {
+ *    ctx: koa variable,
+ *    dataFile: IP storage file path,
+ *    key: config key or request key,
+ *    device: null or request device
+ * }
+ */
 function add(addData) {
   const ctx = addData.ctx;
   const dataFile = addData.dataFile;
@@ -122,7 +164,8 @@ function add(addData) {
   const obj = jsonfile.readFileSync(dataFile, {throws: false});
   const succeededJson = {
     status: "Succeeded",
-    getUrl: config.host + "/?request=get&key=" + key + device
+    getUrl: config.host + "/?request=get&key=" + key + device,
+    deleteUrl: config.host + "/?request=delete&key=" + key + device
   };
   if (obj === null) {
     const obj = {
@@ -146,6 +189,11 @@ function add(addData) {
   };
 };
 
+/**
+ * Get IP
+ * ctx -> koa variable
+ * dataFile -> IP storage file path
+ */
 function get(ctx, dataFile) {
   const obj = jsonfile.readFileSync(dataFile, {throws: false});
   if (obj === null) {
